added tests for parent deactivation
